SmartDocAI/requirements.txt
streamlit==1.28.1
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
azure-search-documents==11.4.0
azure-ai-formrecognizer==3.3.2
azure-storage-blob==12.19.0
openai==1.3.0
python-dotenv==1.0.0
pypdf2==3.0.1
python-docx==0.8.11
pandas==2.1.4
numpy==1.24.3
faiss-cpu==1.7.4
tiktoken==0.5.1
beautifulsoup4==4.12.2
requests==2.31.0
